requirements.txt
pyserial==3.5
pytz==2022.7.1
requests==2.28.2
python-gsmmodem-new==0.13.0
